+++ src/graphics/mod.rs
@@ -1,3 +1,4 @@
 /// NES graphics hardware emulation
 pub mod ppu;
 pub mod ui;
+pub mod palettes;